+++ patches/src/main/kotlin/app/abeja/patches/shared/Constants.kt
@@ -17,15 +17,19 @@ object Constants {
             // It is highly recommended to always include the exact app version you developed your patches for
             // or the last version you have confirmed as 100% working.
             AppTarget(
+                version = null,
+                isExperimental = true
+            ),
+            AppTarget(
                 version = "16.0.002.000.2601041449"
             )
         )
     )
 
 
-    val COMPATIBILITY_EXAMPLE_2 = Compatibility(
-        name = "ABC app",
-        packageName = "com.example.app",
+    val ZTE_VIDEO_PLAYER = Compatibility(
+        name = "ZTE Video Player",
+        packageName = "com.zte.videoplayer",
         apkFileType = ApkFileType.APKM,
         appIconColor = 0x00FF45, // Icon color in Morphe Manager. Usually the same color as the icon background.
         targets = listOf(
@@ -36,7 +40,7 @@ object Constants {
             ),
             // App version confirmed 100% working.
             AppTarget(
-                version = "1.0.2"
+                version = "16.0.000.000.2601071548"
             )
         )
     )