+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.1.0-dev.2](https://github.com/TheRealCrazyfuy/abeja-morphe-patches/compare/v1.1.0-dev.1...v1.1.0-dev.2) (2026-07-14)
+
+### ✨ New Features
+
+* **ZTE Gallery:** Experimental support all versions of ZTE Gallery ([5e01099](https://github.com/TheRealCrazyfuy/abeja-morphe-patches/commit/5e01099da86b4abb3c33918b6a67c0564d005c37))
+
+## [1.1.0-dev.1](https://github.com/TheRealCrazyfuy/abeja-morphe-patches/compare/v1.0.0...v1.1.0-dev.1) (2026-07-14)
+
+### ✨ New Features
+
+* **ZTE Gallery:** Disable AI features ([01f0cea](https://github.com/TheRealCrazyfuy/abeja-morphe-patches/commit/01f0cead0ddaf479e5075a7c87c2dcd702ced186))
+
 ## 1.0.0 (2026-07-14)
 
 ### 🐛 Bug Fixes